Fixing action_unicode
Unicode through action_unicode so Unicode events are not longer sent to
user function.
This has the benefit of making Unicode events to have no chance to
collide with other events in user function (which should be only
keycode and because having unicode processing here is kind of useless).

TODO
Sync and update code to Jack master branch (hex to keycode through
math.h and such).
Add ifdef to stay at 16 bit to reduce footprint when not using full
Unicode support (32 bit).